refactor utility_entry to cli.py
